pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-demisto-sdk"
version = "0.2.0-alpha"
description = "MCP server for demisto-sdk commands with pattern search (RAG)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ciaran Finnegan"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "PyYAML>=6.0",
]
# Note: demisto-sdk must be installed separately (conflicting pydantic versions).
# Install via: pipx install demisto-sdk

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1",
]
rag = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "PyYAML>=6.0",
]

[project.scripts]
mcp-demisto-sdk = "mcp_demisto_sdk.server:main"
xsiam-build-index = "mcp_demisto_sdk.rag.indexer:main"

[project.urls]
Homepage = "https://github.com/ciaran-finnegan/cortex-xsiam-sdk-mcp-tools"
Repository = "https://github.com/ciaran-finnegan/cortex-xsiam-sdk-mcp-tools"
Issues = "https://github.com/ciaran-finnegan/cortex-xsiam-sdk-mcp-tools/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_demisto_sdk*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
strict = true

